Fix intermittent closing of unrelated handles after closeKey.
After a handle has been closed explicitly, it must not be closed again by the finalizer, because the handle value may have been reused for unrelated things in the meantime.

The handle must be invalidated by closeKey, which is not possible with N-API "external" values because they are immutable. Therefore use napi_wrap, for which there is no direct C++ equivalent (Napi::ObjectWrap allocates a separate C++ object rather than treating the HKEY as a pointer and putting it directly into the JS object, as previously done with the external).