Move native code, gypfile to subdir to fix #23.
electron-rebuild will blindly run any top-level binding.gyp, which will fail due to the missing dep, but also is a waste of time as this is N-API and Windows only, so there's no reason to ever rebuild it. Save such users by moving the binding.gyp the only way that works AFAIK, by providing a --cwd to prebuildify.